Update LAB_01_Configuring_Windows_PowerShell_finding_and_running_commands.md
Line 28: Changed "Lab Setup" to "Lab setup".                                               
Line 36: Changed "Lab Startup" to "Lab startup".                                            
Line 92: Changed "Customize the appearance of the ISE to use the single-pane view, hide the **Command** pane, and adjust the font size" to "Customize the ISE's appearance to use a single-pane view, hide the **Command** pane, and adjust the font size".